Implement spell check correction feature for user input countries
If the user input country is off by a single letter, then the game prompts the user with the following message
"Did you mean <correct_spelling> (y/n)". On pressing 'y' keyboard button, the flag of the country is displayed indicating the
correct answer and on pressing any other letter, the usual incorrect message is displayed.